+++ src/connection.ts
@@ -1,19 +1,21 @@
 import { TypedEventEmitter } from '@libp2p/interface'
-import { multiaddr, type Multiaddr } from '@multiformats/multiaddr'
-import { type Uint8ArrayList } from 'uint8arraylist'
+import { multiaddr } from '@multiformats/multiaddr'
 import type * as napi from './napi.js'
 import type { AbortOptions, ComponentLogger, CounterGroup, Direction, Logger, MultiaddrConnection, MultiaddrConnectionTimeline } from '@libp2p/interface'
+import type { Multiaddr } from '@multiformats/multiaddr'
 import type { Sink } from 'it-stream-types'
+import type { Uint8ArrayList } from 'uint8arraylist'
 
 interface QuicConnectionInit {
   connection: napi.Connection
   logger: ComponentLogger
   direction: Direction
   metrics?: CounterGroup
+  metricsPrefix?: string
 }
 
 interface QuicConnectionEvents {
-  'close': CustomEvent
+  close: CustomEvent
 }
 
 export class QuicConnection extends TypedEventEmitter<QuicConnectionEvents> implements MultiaddrConnection {
@@ -22,6 +24,9 @@ export class QuicConnection extends TypedEventEmitter<QuicConnectionEvents> impl
   readonly log: Logger
   readonly remoteAddr: Multiaddr
   readonly metrics?: CounterGroup
+  readonly metricsPrefix: string
+
+  private remoteClosed?: boolean
 
   timeline: MultiaddrConnectionTimeline = {
     open: Date.now()
@@ -37,9 +42,12 @@ export class QuicConnection extends TypedEventEmitter<QuicConnectionEvents> impl
     this.log = init.logger.forComponent(`libp2p:quic:connection:${this.#connection.id()}:${init.direction}`)
     this.remoteAddr = multiaddr(this.#connection.remoteMultiaddr())
     this.metrics = init.metrics
+    this.metricsPrefix = init.metricsPrefix ?? ''
 
     // close maconn when connection is closed by remote
     this.#connection.closed().then(() => {
+      this.remoteClosed = true
+      this.metrics?.increment({ [`${this.metricsPrefix}end`]: true })
       this.close()
         .catch(err => {
           this.abort(err)
@@ -58,7 +66,11 @@ export class QuicConnection extends TypedEventEmitter<QuicConnectionEvents> impl
 
     this.timeline.close = Date.now()
     this.log('closed')
-    this.metrics?.increment({ close: true })
+
+    if (this.remoteClosed !== true) {
+      this.metrics?.increment({ [`${this.metricsPrefix}close`]: true })
+    }
+
     this.safeDispatchEvent('close')
   }
 
@@ -71,7 +83,7 @@ export class QuicConnection extends TypedEventEmitter<QuicConnectionEvents> impl
 
     this.timeline.close = Date.now()
     this.log('aborted - %e', err)
-    this.metrics?.increment({ abort: true })
+    this.metrics?.increment({ [`${this.metricsPrefix}abort`]: true })
     this.safeDispatchEvent('close')
   }
 }