+++ src/transport.ts
@@ -2,15 +2,17 @@ import { privateKeyToProtobuf } from '@libp2p/crypto/keys'
 import { AbortError, serviceCapabilities, transportSymbol } from '@libp2p/interface'
 import { QuicConnection } from './connection.js'
 import { dialFilter, listenFilter } from './filter.js'
-import { QuicListener, type QuicCreateListenerOptions } from './listener.js'
+import { QuicListener } from './listener.js'
 import * as napi from './napi.js'
 import { QuicStreamMuxerFactory } from './stream-muxer.js'
+import type { QuicComponents, QuicDialOptions, QuicOptions } from './index.js'
+import type { QuicCreateListenerOptions } from './listener.js'
 import type { Connection, CounterGroup, Listener, Logger, MultiaddrFilter, Transport } from '@libp2p/interface'
 import type { Multiaddr } from '@multiformats/multiaddr'
-import type { QuicComponents, QuicDialOptions, QuicOptions } from './index.js'
 
 interface QuicTransportMetrics {
-  events: CounterGroup
+  events?: CounterGroup
+  errors?: CounterGroup
 }
 
 export class QuicTransport implements Transport {
@@ -22,7 +24,7 @@ export class QuicTransport implements Transport {
 
   readonly log: Logger
   readonly components: QuicComponents
-  readonly metrics?: QuicTransportMetrics
+  readonly metrics: QuicTransportMetrics
 
   readonly #config: napi.QuinnConfig
 
@@ -47,13 +49,15 @@ export class QuicTransport implements Transport {
       ip6: new napi.Client(this.#config, 1)
     }
 
-    if (this.components.metrics != null) {
-      this.metrics = {
-        events: this.components.metrics?.registerCounterGroup('libp2p_quic_dialer_events_total', {
-          label: 'event',
-          help: 'Total count of QUIC dialer events by type'
-        })
-      }
+    this.metrics = {
+      events: this.components.metrics?.registerCounterGroup('libp2p_quic_dialer_events_total', {
+        label: 'event',
+        help: 'Total count of QUIC dialer events by type'
+      }),
+      errors: this.components.metrics?.registerCounterGroup('libp2p_quic_dialer_errors_total', {
+        label: 'event',
+        help: 'Total count of QUIC dialer errors by type'
+      })
     }
 
     this.listenFilter = listenFilter
@@ -73,25 +77,41 @@ export class QuicTransport implements Transport {
 
     const dialPromise = dialer.outboundConnection(addr.address, addr.port)
     dialPromise
-      .then(() => this.metrics?.events.increment({ connect: true }))
-      .catch(() => this.metrics?.events.increment({ error: true }))
+      .then(() => this.metrics.events?.increment({ connect: true }))
+      .catch(() => this.metrics.events?.increment({ error: true }))
     const connection = await dialPromise
 
-    const maConn = new QuicConnection({
-      connection,
-      logger: this.components.logger,
-      direction: 'outbound',
-      metrics: this.metrics?.events
-    })
-    return options.upgrader.upgradeOutbound(maConn, {
-      skipEncryption: true,
-      skipProtection: true,
-      muxerFactory: new QuicStreamMuxerFactory({
+    let maConn: QuicConnection
+
+    try {
+      maConn = new QuicConnection({
         connection,
-        logger: this.components.logger
-      }),
-      signal: options.signal
-    })
+        logger: this.components.logger,
+        direction: 'outbound',
+        metrics: this.metrics?.events
+      })
+    } catch (err) {
+      this.metrics.errors?.increment({ outbound_to_connection: true })
+      throw err
+    }
+
+    try {
+      this.log('new outbound connection %a', maConn.remoteAddr)
+      return await options.upgrader.upgradeOutbound(maConn, {
+        skipEncryption: true,
+        skipProtection: true,
+        muxerFactory: new QuicStreamMuxerFactory({
+          connection,
+          logger: this.components.logger
+        }),
+        signal: options.signal
+      })
+    } catch (err: any) {
+      this.metrics.errors?.increment({ outbound_upgrade: true })
+      this.log.error('error upgrading outbound connection - %e', err)
+      maConn.abort(err)
+      throw err
+    }
   }
 
   createListener (options: QuicCreateListenerOptions): Listener {